fixed producttask

+++ src/main/java/bgu/spl/a2/sim/GsonReader.java
@@ -20,6 +20,6 @@ public class GsonReader {
 	class Zerg {
 		String product;
 		int qty;
-		int startId;
+		long startId;
 	}
 }
